+++ Main/ClientApp/src/Model/SharedCollection.ts
@@ -1,0 +1,6 @@
+export type SharedCollection = {
+  collectionId: number;
+  userId: string;
+  viewRights: boolean;
+  editRights: boolean;
+};
